Uploaded project to GitHub.
Initial status of this commit.
The 5s is mostly working.
The random number selection is working.
The slider is working and resets when you start a new round.
The Hit Me! button works.
The popup alert is working.
The Score goes up per points.
The Round goes up with every new round.
info button takes you to the next screen.
The close button works and returns to the first screen.
The restart button gives you a new random number as it should.
The restart button does not reset the Score.
The restart button increases the number of the Round when you click it. It shouldn't do this.
The restart button should reset the Round to 1.

For the 6sPlus, the same things are working in the first screen
The second screen, the only thing that comes up is the close button and it is all the way in the upper left corner.
The Close button does work and closes the screen.